refactor(clean): split css.c into topical headers
css.c had grown to 4600 lines, past the size where one file stays
navigable. It now includes seven topical fragments (tokenize, value, calc,
render, shorthand, selector, grammar), each under 800 lines, in the order
the code already read top to bottom. The fragments compile as one
translation unit, the way tokenizer/statemachine_run.h does, so every static
helper stays inline and the generated code is unchanged; its speed held
within 0.2% across an interleaved A/B, and per-file coverage stays 100%.
